+++ tests/Mocks/StorageEngineMock.h
@@ -81,7 +81,7 @@ class PhysicalCollectionMock : public arangodb::PhysicalCollection {
   PhysicalCollectionMock(arangodb::LogicalCollection& collection);
   std::shared_ptr<arangodb::Index> createIndex(
       arangodb::velocypack::Slice info, bool restore, bool& created,
-      std::shared_ptr<std::function<arangodb::Result(uint64_t)>> =
+      std::shared_ptr<std::function<arangodb::Result(double)>> =
           nullptr) override;
   void deferDropCollection(
       std::function<bool(arangodb::LogicalCollection&)> const& callback)
